// src/views/baseConfig/taskType/utils/types.ts
export interface FormItemProps {
  task_type_id?: number;
  task_type_name?: string;
  task_type_code?: string;
  task_type_desc?: string;
  is_other_operation?: number;
  status?: number;
  display?: number;
  sort?: number;
  is_use?: number;
  is_def?: number;
  tracker_button?: string;
  is_record_end_time?: number;
  can_transfer?: number;
  is_merge_tube?: number;
  is_countdown?: number;
  create_by?: number;
  created_at?: string;
  update_by?: number;
  updated_at?: string;
}